# backend/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: https://localhost:8443
    description: Generated server url
...
paths:
  /api/v1/profile/{id}:
    get:
      tags:
        - Profile API
      summary: Get profile
      description: Returns complete profile data for a given user id.
      operationId: profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/ProfileDTO"
    put:
      tags:
        - Profile API
      summary: Update profile (JSON)
      description: Updates profile data from JSON payload for the authenticated owner.
      operationId: editProfileMultipart
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                request:
                  $ref: "#/components/schemas/ProfileUpdateRequestDTO"
                imageFile:
                  type: string
                  format: binary
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdateRequestDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
  /api/v1/events/{id}:
    get:
      tags:
        - Event API
      summary: Get event detail
      description:
        Returns a full event detail payload including management and purchase
        flags.
      operationId: getEvent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
                additionalProperties:
                  type: object
    put:
      tags:
        - Event API
      summary: Update event
      description: Updates an event if the current user is authorized.
      operationId: updateEvent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: tagNames
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                event:
                  $ref: "#/components/schemas/EventDTO"
                imageFile:
                  type: string
                  format: binary
              required:
                - event
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
    delete:
      tags:
        - Event API
      summary: Delete event
      description: Deletes an event if the current user is authorized.
      operationId: deleteEvent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
...
  /api/v1/events/:
    get:
      tags:
        - Event API
      summary: List events
      description:
        Returns paginated published events with optional search and tag
        filters.
      operationId: getEvents
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: tags
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  type: object
                  additionalProperties:
                    type: object
    post:
      tags:
        - Event API
      summary: Create event
      description:
        "Creates an event from multipart data, optionally including image\
        \ and tags."
      operationId: createEvent
      parameters:
        - name: tagNames
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                event:
                  $ref: "#/components/schemas/EventDTO"
                imageFile:
                  type: string
                  format: binary
              required:
                - event
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
...
  /api/v1/cart/add/event/{id}:
    post:
      tags:
        - Cart API
      summary: Add event to cart
      description: Adds an event to the authenticated user's pending order.
      operationId: addEventToCart
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
...
  /api/v1/auth/refresh:
    post:
      tags:
        - Authentication API
      summary: Refresh token
      description:
        Issues fresh credentials using a refresh token from cookie or authorization
        header.
      operationId: refresh
      parameters:
        - name: RefreshToken
          in: cookie
          required: false
          schema:
            type: string
        - name: Authorization
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /api/v1/images/events/{eventId}:
    get:
      tags:
        - Image API
      summary: Get event image
      description: Returns event image bytes for an event id.
      operationId: getEventImage
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: string
                format: byte
...
  /api/v1/events/location-search:
    get:
      tags:
        - Event API
      summary: Search locations
      description: Searches event locations using the external geocoding service.
      operationId: searchLocations
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
...
  /api/v1/admin/events:
    get:
      tags:
        - Admin API
      summary: List events
      description:
        Returns paginated events for admin management with optional text
        filtering.
      operationId: getEvents_1
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventDTO"
...
components:
  schemas:
    ProfileUpdateRequestDTO:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        country:
          type: string
        shortDescription:
          type: string
        currentGoal:
          type: string
        weeklyRoutine:
          type: string
        comunity:
          type: string
    EventDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        priceCents:
          type: integer
          format: int32
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        capacity:
          type: integer
          format: int32
        category:
          type: string
        status:
          type: string
          enum:
            - DRAFT
            - PENDING_REVIEW
            - PUBLISHED
            - ARCHIVED
...
    AuthResponse:
      type: object
      properties:
        status:
          type: string
          enum:
            - SUCCESS
            - FAILURE
        message:
          type: string
        error:
          type: string
...
    ProfileDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        country:
          type: string
        shortDescription:
          type: string
        currentGoal:
          type: string
        weeklyRoutine:
          type: string
        comunity:
          type: string
        profileImage:
          type: string
        isProfileOwner:
          type: boolean
        userType:
          type: string
        completedCourses:
          type: integer
          format: int32
        completedCourseNames:
          type: array
          items:
            type: string
        inProgressCount:
          type: integer
          format: int32
        userTags:
          type: array
          items:
            type: string
          uniqueItems: true
        subscribedCourses:
          type: array
          items:
            type: object
            additionalProperties:
              type: object
        userEvents:
          type: array
          items:
            type: object
            additionalProperties:
              type: object
        averageProgress:
          type: integer
          format: int32
        totalEnrollments:
          type: integer
          format: int32
        totalLessonsCompleted:
          type: integer
          format: int64
        completedLessonsThisMonth:
          type: integer
          format: int64
        averageLessonsPerMonth:
          type: string
        createdCourses:
          type: array
          items:
            type: object
            additionalProperties:
              type: object
        hasMultipleCourses:
          type: boolean
...