# backend/scam-g18/src/main/java/es/codeurjc/scam_g18/controller/ProfileWebController.java
package es.codeurjc.scam_g18.controller;

import java.io.IOException;
import java.security.Principal;
import java.sql.SQLException;
import java.util.Optional;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.Authentication;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.multipart.MultipartFile;

import es.codeurjc.scam_g18.model.User;
import es.codeurjc.scam_g18.service.CourseService;
import es.codeurjc.scam_g18.service.EnrollmentService;
import es.codeurjc.scam_g18.service.UserService;
import jakarta.servlet.http.HttpServletRequest;

@Controller
public class ProfileWebController {

    @Autowired
    private UserService userService;

    @Autowired
    private EnrollmentService enrollmentService;

    @Autowired
    private CourseService courseService;

    @GetMapping("/profile/me")
    public String myProfile(Principal principal) {
        if (principal == null)
            return "redirect:/login";
        return userService.getCurrentAuthenticatedUser()
                .map(user -> "redirect:/profile/" + user.getId())
                .orElse("redirect:/login");
    }

    @GetMapping("/profile/{id}")
    public String profile(Model model, @PathVariable long id) {
        Optional<User> user = userService.findById(id);
        if (user.isEmpty())
            return "redirect:/";

        var currentUserOpt = userService.getCurrentAuthenticatedUser();
        boolean isProfileOwner = currentUserOpt.isPresent() && currentUserOpt.get().getId().equals(id);

        model.addAttribute("user", user.get());
        model.addAttribute("isProfileOwner", isProfileOwner);
        model.addAttribute("profileImage", userService.getProfileImage(id));
        model.addAttribute("completedCourses", userService.getCompletedCoursesCount(id));
        model.addAttribute("completedCourseNames", enrollmentService.getCompletedCourseNames(id));
        model.addAttribute("inProgressCount", enrollmentService.getInProgressCount(id));
        model.addAttribute("userType", userService.getUserType(id));
        model.addAttribute("userTags", enrollmentService.getTagNamesByUserId(id));
        model.addAttribute("subscribedCourses", enrollmentService.getSubscribedCoursesData(id));
        model.addAttribute("userEvents", enrollmentService.getUserEvents(id));
        model.addAttribute("averageProgress", enrollmentService.getAverageProgress(id));
        model.addAttribute("totalEnrollments", enrollmentService.getTotalEnrollments(id));
        model.addAttribute("totalLessonsCompleted", enrollmentService.getTotalCompletedLessons(id));
        model.addAttribute("completedLessonsThisMonth", enrollmentService.getLessonsCompletedThisMonth(id));
        model.addAttribute("averageLessonsPerMonth", enrollmentService.getAverageLessonsPerMonthFormatted(id));

        // Retrieve created courses explicitly for publisher panel stats
        java.util.List<java.util.Map<String, Object>> createdCourses = courseService.getCreatedCoursesWithStats(id);
        model.addAttribute("createdCourses", createdCourses);
        model.addAttribute("hasMultipleCourses", createdCourses.size() > 1);

        model.addAttribute("listaPaises", userService.getAllCountries());

        return "profile";
    }

    @PostMapping("/profile/{id}/edit")
    public String editProfile(@PathVariable long id,
            @RequestParam String username,
            @RequestParam String email,
            @RequestParam(required = false) String country,
            @RequestParam(required = false) String shortDescription,
            @RequestParam(required = false) String currentGoal,
            @RequestParam(required = false) String weeklyRoutine,
            @RequestParam(required = false) String comunity,
            @RequestParam(required = false) MultipartFile imageFile,
            HttpServletRequest request,
            org.springframework.web.servlet.mvc.support.RedirectAttributes redirectAttributes)
            throws IOException, SQLException {

        var currentUserOpt = userService.getCurrentAuthenticatedUser();
        if (currentUserOpt.isEmpty() || !currentUserOpt.get().getId().equals(id)) {
            return "redirect:/login";
        }

        // Use the updated UserService validator. We pass nulls to
        // password/gender/birthDate since they aren't edited here.
        String validationErrors = userService.validateUserAttributes(username, email, null, null, null, null);
        if (validationErrors != null) {
            redirectAttributes.addFlashAttribute("errorMessage", validationErrors);
            return "redirect:/profile/" + id;
        }

        userService.updateProfile(id, username, email, country, shortDescription, currentGoal, weeklyRoutine,
                comunity, imageFile);

        // Refresh the Spring Security session if the username changed
        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
        if (auth != null && username != null && !username.isBlank() && !auth.getName().equals(username)) {
            userService.refreshUserSession(username, request);
        }

        return "redirect:/profile/" + id;
    }

}